Refactor component definition locations for API component
